requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
bpy==3.6.0
bracex==2.4
certifi==2023.7.22
charset-normalizer==3.3.2
cli-exit-tools==1.2.6
click==8.1.7
colorama==0.4.6
Cython==3.0.5
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
lib-detect-testenv==2.0.8
numpy==1.26.1
OpenEXR==1.3.9
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
watchfiles==0.21.0
wcmatch==8.5
websockets==12.0
zstandard==0.22.0
